Granularize vector's unit tests
